requirements.txt
fastapi
uvicorn
pandas
yfinance
ta
python-telegram-bot

